Refactor RF4SWebServer
Refactor `RF4SWebServer.ts` into smaller files to improve maintainability.
